+++ src/main/java/com/gymapp/controller/UsuarioController.java
@@ -8,23 +8,24 @@ import org.springframework.http.ResponseEntity;
 import java.util.List;
 
 public class UsuarioController implements UsuariosApi {
+
     @Override
     public ResponseEntity<Void> usuariosPost(Usuario usuario) {
         return null;
     }
 
     @Override
-    public ResponseEntity<Usuario> usuariosUsuarioIdGet(Integer usuarioId) {
+    public ResponseEntity<Usuario> usuariosUsuarioIdGet(Long usuarioId) {
         return null;
     }
 
     @Override
-    public ResponseEntity<List<Programa>> usuariosUsuarioIdProgramaGet(Integer usuarioId) {
+    public ResponseEntity<List<Programa>> usuariosUsuarioIdProgramaGet(Long usuarioId) {
         return null;
     }
 
     @Override
-    public ResponseEntity<Void> usuariosUsuarioIdPut(Integer usuarioId, Usuario usuario) {
+    public ResponseEntity<Void> usuariosUsuarioIdPut(Long usuarioId, Usuario usuario) {
         return null;
     }
 }
